Support undefined type in IDL parser.
This change was inflicted on us in heycam/webidl#906 It is quite possible downstream
consumers will refuse to adopt this change as it obliterates the difference between
IDL definition and js binding and breaks almost every WebIDL in existance.

We retain support for void promises/return types to retain compatibility with the
existing ecosystem.