Add @HasNoSideEffects/@nosideeffects to generated java/closure artifacts
Add the `@HasNoSideEffects` annotation to selected operations in the
akasha generated jsinterop classes. Add the equivalent annotation
`@nosideeffects` to the generated closure externs. The generation was
driven by the presence of the `[NoSideEffects]` extended attribute on
the operation. The `[NoSideEffects]` extended attribute was added to
the WebIDL via a processor that looked in the file
`data/pipelines/no_side_effects.txt` to identify operations that should
have this extended attribute. The list of operations was sourced from
the existing closure externs. Additional heuristics were used to add
`NoSideEffects` such as any named getters as well as the read
operations defined by `maplike` WebIDL construct.

Fixes #13.